+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- **Fix** : Fixed PHP 8 issue
+
 ## 2.3.1
 
 - **Fix**: Fixed 3DS v2 merchant risk information builder