+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- **Add** : Sisal reference without redirection
 - **Fix** : Fixed PHP 8 issue
 
 ## 2.3.1