+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- **Add** : Sisal reference without redirection
+- **Add** : Added Sisal reference without redirection
+- **Add** : Added Mooney reference without redirection
 - **Fix** : Fixed PHP 8 issue
 
 ## 2.3.1